+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
+## [1.1.0] - 2023-04-24
+### Added
+ - poetry run cli -> scripts/cli.py
+ - for extra dependencies install with `poetry install --with scripts`
+ - get_ruptures query
+ - get_rupture_ids query
+
+### Changed
+ - requires nzshm-model >=0.3.0
+
 ## [1.0.1] - 2023-04-19
-## Removed
+### Removed
  - geopandas dependency
 
 ## [1.0.0] - 2023-04-19